Implement manager functions and add user delete check

+++ frontend/src/app/+user/user/user.component.ts
@@ -26,6 +26,8 @@ import { SessionQuery } from "../../store/session.query";
 import { MatSnackBar } from "@angular/material/snack-bar";
 import { EditManagerDialogComponent } from "./edit-manager-dialog/edit-manager-dialog.component";
 import { ConfigService } from "../../services/config/config.service";
+import { ErrorDialogComponent } from "../../dialogs/error/error-dialog.component";
+import { IgeError } from "../../models/ige-error";
 
 @UntilDestroy()
 @Component({
@@ -144,22 +146,35 @@ export class UserComponent
   }
 
   deleteUser(login: string) {
-    this.dialog
-      .open(ConfirmDialogComponent, {
-        data: {
-          title: "Benutzer löschen",
-          message: "Möchten Sie den Benutzer " + login + " wirklich löschen?",
-        } as ConfirmDialogData,
-      })
-      .afterClosed()
-      .subscribe((result) => {
-        if (result) {
-          this.userService.deleteUser(login).subscribe(() => {
-            this.selectedUser$.next(null);
-            this.fetchUsers();
+    this.userService.getManagedUsers(login).subscribe((managedUsers) => {
+      if (managedUsers?.length) {
+        this.dialog.open(ErrorDialogComponent, {
+          data: new IgeError(
+            "Der Benutzer kann nicht gelöscht werden, da er noch für folgende Nutzer verantwortlich ist:\n\n" +
+              managedUsers.join("\n") +
+              "\n\nBitte ändern Sie zunächst den Verantwortlichen der betroffenen Benutzer."
+          ),
+        });
+      } else {
+        this.dialog
+          .open(ConfirmDialogComponent, {
+            data: {
+              title: "Benutzer löschen",
+              message:
+                "Möchten Sie den Benutzer " + login + " wirklich löschen?",
+            } as ConfirmDialogData,
+          })
+          .afterClosed()
+          .subscribe((result) => {
+            if (result) {
+              this.userService.deleteUser(login).subscribe(() => {
+                this.selectedUser$.next(null);
+                this.fetchUsers();
+              });
+            }
           });
-        }
-      });
+      }
+    });
   }
 
   saveUser(user?: User): void {
